Modularize course features into separate Kotlin files
Moves core data models (e.g., CourseItem, CourseCategory) and various functional blocks (e.g., attachment handling, progress tracking, download management, UI actions) out of CourseWizardActivity and DashboardCoursePageFragment. This improves code organization, reduces class complexity, and facilitates easier maintenance.
