+++ app/src/main/java/org/ole/planet/myplanet/lite/CourseWizardActivity.kt
@@ -9,50 +9,33 @@ import android.content.Context
 import android.content.Intent
 import android.os.Bundle
 import android.view.View
-import android.widget.ImageButton
 import android.widget.LinearLayout
 import android.widget.TextView
 import android.widget.Toast
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.core.content.IntentCompat
-import androidx.core.content.edit
-import androidx.core.net.toUri
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowCompat
 import androidx.core.view.WindowInsetsCompat
 import androidx.core.view.WindowInsetsControllerCompat
 import androidx.lifecycle.lifecycleScope
-import androidx.media3.common.MediaItem
-import androidx.media3.datasource.DataSource
-import androidx.media3.datasource.DefaultDataSource
-import androidx.media3.datasource.DefaultHttpDataSource
 import androidx.media3.exoplayer.ExoPlayer
-import androidx.media3.exoplayer.source.DefaultMediaSourceFactory
-import androidx.media3.ui.PlayerView
 import com.google.android.material.snackbar.Snackbar
 import io.noties.markwon.Markwon
 import io.noties.markwon.ext.tables.TablePlugin
 import io.noties.markwon.html.HtmlPlugin
 import io.noties.markwon.image.glide.GlideImagesPlugin
-import java.util.Locale
 import kotlinx.coroutines.launch
-import okhttp3.Credentials
-import org.json.JSONArray
-import org.ole.planet.myplanet.lite.BaseActivity
 import org.ole.planet.myplanet.lite.dashboard.DashboardCoursesRepository
-import org.ole.planet.myplanet.lite.dashboard.DashboardImagePreviewActivity
 import org.ole.planet.myplanet.lite.dashboard.DashboardServerPreferences
-import org.ole.planet.myplanet.lite.dashboard.DashboardSurveysRepository.SurveyDocument
 import org.ole.planet.myplanet.lite.profile.ProfileCredentialsStore
 import org.ole.planet.myplanet.lite.profile.StoredCredentials
 import org.ole.planet.myplanet.lite.survey.DashboardLocalSurveyRepository
-import org.ole.planet.myplanet.lite.util.MarkdownUtils
-import org.ole.planet.myplanet.lite.util.NetworkUtils
 import org.ole.planet.myplanet.lite.util.SecurePreferencesProvider
 
 @androidx.annotation.OptIn(androidx.media3.common.util.UnstableApi::class)
 class CourseWizardActivity : BaseActivity() {
-    private val pendingProgressPrefs by lazy {
+    val pendingProgressPrefs by lazy {
         SecurePreferencesProvider.getEncryptedPreferences(
             context = applicationContext,
             prefsName = PREF_ENCRYPTED_PENDING_COURSE_PROGRESS,
@@ -61,11 +44,11 @@ class CourseWizardActivity : BaseActivity() {
     }
 
     private lateinit var markwon: Markwon
-    private var steps: List<StepDisplay> = emptyList()
-    private var baseUrl: String? = null
-    private var currentIndex: Int = 0
-    private var courseId: String? = null
-    private var credentials: StoredCredentials? = null
+    var steps: List<StepDisplay> = emptyList()
+    var baseUrl: String? = null
+    var currentIndex: Int = 0
+    var courseId: String? = null
+    var credentials: StoredCredentials? = null
     private lateinit var stepPositionView: TextView
     private lateinit var stepTitleView: TextView
     private lateinit var descriptionView: TextView
@@ -74,19 +57,19 @@ class CourseWizardActivity : BaseActivity() {
     private lateinit var attachmentsTitle: TextView
     private lateinit var previousButton: View
     private lateinit var nextButton: View
-    private val playlistIndexByResourceId = mutableMapOf<String, Int>()
-    private val currentPlaylistUrls = mutableListOf<String>()
-    private var lastPlaybackIndex = 0
-    private var lastPlaybackPositionMs = 0L
-    private val coursesRepository = DashboardCoursesRepository()
-    private val localSurveyRepository by lazy { DashboardLocalSurveyRepository(applicationContext) }
-    private var hasAutoCompletedFirstStep = false
-    private val audioPlayers = mutableListOf<ExoPlayer>()
+    val playlistIndexByResourceId = mutableMapOf<String, Int>()
+    val currentPlaylistUrls = mutableListOf<String>()
+    var lastPlaybackIndex = 0
+    var lastPlaybackPositionMs = 0L
+    val coursesRepository = DashboardCoursesRepository()
+    val localSurveyRepository by lazy { DashboardLocalSurveyRepository(applicationContext) }
+    var hasAutoCompletedFirstStep = false
+    val audioPlayers = mutableListOf<ExoPlayer>()
     private val completedRequiredSteps = mutableSetOf<Int>()
-    private var pendingRequiredStepIndex: Int? = null
-    private var cachedProgressDocument: DashboardCoursesRepository.CourseProgressDocument? = null
+    var pendingRequiredStepIndex: Int? = null
+    var cachedProgressDocument: DashboardCoursesRepository.CourseProgressDocument? = null
 
-    private val fullscreenLauncher =
+    val fullscreenLauncher =
         registerForActivityResult(ActivityResultContracts.StartActivityForResult()) { result ->
             val data = result.data ?: return@registerForActivityResult
             lastPlaybackIndex = data.getIntExtra(FullscreenPlayerActivity.EXTRA_RESULT_INDEX, 0)
@@ -96,7 +79,7 @@ class CourseWizardActivity : BaseActivity() {
             )
         }
 
-    private val requiredStepLauncher =
+    val requiredStepLauncher =
         registerForActivityResult(ActivityResultContracts.StartActivityForResult()) { result ->
             if (result.resultCode == RESULT_OK) {
                 pendingRequiredStepIndex?.let { completedRequiredSteps.add(it) }
@@ -147,7 +130,7 @@ class CourseWizardActivity : BaseActivity() {
             EXTRA_STEPS,
             ArrayList::class.java
         )
-        val stepPayload = rawSteps?.filterIsInstance<DashboardCoursePageFragment.CourseItem.LessonStep>()
+        val stepPayload = rawSteps?.filterIsInstance<CourseItem.LessonStep>()
         steps = stepPayload?.map { step ->
             StepDisplay(
                 title = step.title,
@@ -202,36 +185,7 @@ class CourseWizardActivity : BaseActivity() {
         )
     }
 
-    private suspend fun resolveInitialStepIndex(fallbackIndex: Int): Int {
-        if (cachedProgressDocument != null) {
-            val progressStep = cachedProgressDocument?.stepNum
-            val resolvedIndex = progressStep?.minus(1) ?: fallbackIndex
-            return resolvedIndex.coerceIn(0, steps.lastIndex)
-        }
-        val normalizedBase = baseUrl?.trim()?.trimEnd('/')?.takeIf { it.isNotEmpty() } ?: return fallbackIndex
-        val creds = credentials ?: return fallbackIndex
-        val id = courseId?.takeIf { it.isNotBlank() } ?: return fallbackIndex
-        getPendingProgress(id).maxOrNull()
-        val progressDocuments = coursesRepository.fetchCoursesProgressDocuments(
-            normalizedBase,
-            creds,
-            listOf(id)
-        ).getOrNull()
-        cachedProgressDocument = progressDocuments?.get(id)
-        val progressStep = cachedProgressDocument?.stepNum
-        val resolvedIndex = progressStep?.minus(1) ?: fallbackIndex
-        return resolvedIndex.coerceIn(0, steps.lastIndex)
-    }
-
-    private fun maybeAutoCompleteFirstStep() {
-        if (hasAutoCompletedFirstStep || currentIndex != 0) return
-        hasAutoCompletedFirstStep = true
-        lifecycleScope.launch {
-            runCatching { updateCourseProgressIfNeeded(1) }
-        }
-    }
-
-    private fun bindStep() {
+    fun bindStep() {
         val step = steps[currentIndex]
         stepPositionView.text = getString(
             R.string.course_wizard_step_position,
@@ -301,594 +255,6 @@ class CourseWizardActivity : BaseActivity() {
         ).show()
     }
 
-    private fun advanceToNextStep() {
-        val targetIndex = (currentIndex + 1).coerceAtMost(steps.lastIndex)
-        val targetStepNumber = targetIndex + 1
-        lifecycleScope.launch {
-            runCatching { updateCourseProgressIfNeeded(targetStepNumber) }
-        }
-        currentIndex = targetIndex
-        bindStep()
-    }
-
-    private suspend fun updateCourseProgressIfNeeded(targetStepNumber: Int) {
-        val normalizedBase = baseUrl?.trim()?.trimEnd('/')?.takeIf { it.isNotEmpty() } ?: return
-        val creds = credentials ?: return
-        val id = courseId?.takeIf { it.isNotBlank() } ?: return
-        if (!NetworkUtils.isDeviceOnline(this)) {
-            enqueuePendingProgress(id, targetStepNumber)
-            return
-        }
-        val existingDocuments = coursesRepository.fetchCoursesProgressDocuments(normalizedBase, creds, listOf(id))
-            .getOrNull()
-        val existingDoc = existingDocuments?.get(id)
-        val existingStep = existingDoc?.stepNum ?: 0
-        if (existingStep >= targetStepNumber) return
-        val progressDoc = existingDoc ?: cachedProgressDocument
-        val now = System.currentTimeMillis()
-        val document = DashboardCoursesRepository.CourseProgressUpdateDocument(
-            id = progressDoc?.id,
-            rev = progressDoc?.rev,
-            userId = "org.couchdb.user:${creds.username}",
-            courseId = id,
-            stepNum = targetStepNumber,
-            passed = true,
-            createdOn = progressDoc?.createdOn
-                ?: DashboardServerPreferences.getServerCode(applicationContext),
-            parentCode = progressDoc?.parentCode
-                ?: DashboardServerPreferences.getServerParentCode(applicationContext),
-            createdDate = progressDoc?.createdDate ?: now,
-            updatedDate = now
-        )
-        val saveResult = coursesRepository.saveCourseProgress(normalizedBase, creds, listOf(document))
-        if (saveResult.isFailure) {
-            enqueuePendingProgress(id, targetStepNumber)
-        } else {
-            val persistedDoc = saveResult.getOrNull()
-                ?.firstOrNull { it.ok == true || (!it.id.isNullOrBlank() && !it.rev.isNullOrBlank()) }
-            val resolvedId = persistedDoc?.id ?: document.id
-            val resolvedRev = persistedDoc?.rev ?: document.rev
-            cachedProgressDocument = DashboardCoursesRepository.CourseProgressDocument(
-                id = resolvedId,
-                rev = resolvedRev,
-                courseId = document.courseId,
-                stepNum = document.stepNum,
-                passed = document.passed,
-                createdDate = document.createdDate,
-                updatedDate = document.updatedDate,
-                createdOn = document.createdOn,
-                parentCode = document.parentCode
-            )
-        }
-    }
-
-    private suspend fun finishCourse() {
-        runCatching { updateCourseProgressIfNeeded(steps.size) }
-        setResult(
-            RESULT_OK,
-            Intent().apply {
-                putExtra(EXTRA_RESULT_COURSE_ID, courseId)
-                putExtra(EXTRA_RESULT_PROGRESS_PERCENT, 100)
-                putExtra(EXTRA_RESULT_CURRENT_STEP, steps.lastIndex)
-            }
-        )
-        finish()
-    }
-
-    private suspend fun flushPendingCourseProgress() {
-        val normalizedBase = baseUrl?.trim()?.trimEnd('/')?.takeIf { it.isNotEmpty() } ?: return
-        val creds = credentials ?: return
-        val id = courseId?.takeIf { it.isNotBlank() } ?: return
-        if (!NetworkUtils.isDeviceOnline(this)) return
-        val pendingSteps = getPendingProgress(id)
-        if (pendingSteps.isEmpty()) return
-
-        val existingDoc = coursesRepository.fetchCoursesProgressDocuments(normalizedBase, creds, listOf(id))
-            .getOrNull()?.get(id)
-        val existingStep = existingDoc?.stepNum ?: 0
-        val maxPendingStep = pendingSteps.maxOrNull() ?: 0
-
-        if (maxPendingStep <= existingStep) {
-            removePendingProgressBatch(id, pendingSteps.toSet())
-            return
-        }
-
-        val now = System.currentTimeMillis()
-        val document = DashboardCoursesRepository.CourseProgressUpdateDocument(
-            id = existingDoc?.id,
-            rev = existingDoc?.rev,
-            userId = "org.couchdb.user:${creds.username}",
-            courseId = id,
-            stepNum = maxPendingStep,
-            passed = true,
-            createdOn = existingDoc?.createdOn
-                ?: DashboardServerPreferences.getServerCode(applicationContext),
-            parentCode = existingDoc?.parentCode
-                ?: DashboardServerPreferences.getServerParentCode(applicationContext),
-            createdDate = existingDoc?.createdDate ?: now,
-            updatedDate = now
-        )
-
-        val result = coursesRepository.saveCourseProgress(normalizedBase, creds, listOf(document))
-        if (result.isSuccess) {
-            removePendingProgressBatch(id, pendingSteps.toSet())
-            val persistedDoc = result.getOrNull()
-                ?.firstOrNull { it.ok == true || (!it.id.isNullOrBlank() && !it.rev.isNullOrBlank()) }
-            val resolvedId = persistedDoc?.id ?: existingDoc?.id
-            val resolvedRev = persistedDoc?.rev ?: existingDoc?.rev
-            cachedProgressDocument = DashboardCoursesRepository.CourseProgressDocument(
-                id = resolvedId,
-                rev = resolvedRev,
-                courseId = id,
-                stepNum = maxPendingStep,
-                passed = true,
-                createdDate = document.createdDate,
-                updatedDate = now,
-                createdOn = document.createdOn,
-                parentCode = document.parentCode
-            )
-        }
-    }
-
-    private suspend fun flushPendingExamSubmissions() {
-        localSurveyRepository.flushPendingSurveyOutbox("exam")
-    }
-
-
-
-    private fun enqueuePendingProgress(courseId: String, stepNumber: Int) {
-        val updated = (getPendingProgress(courseId) + stepNumber).distinct().sorted()
-        val array = JSONArray()
-        updated.forEach { array.put(it) }
-        pendingProgressPrefs.edit { putString(progressKey(courseId), array.toString()) }
-    }
-
-    private fun getPendingProgress(courseId: String): List<Int> {
-        val raw = pendingProgressPrefs.getString(progressKey(courseId), null) ?: return emptyList()
-        return runCatching {
-            val array = JSONArray(raw)
-            (0 until array.length()).mapNotNull { index ->
-                val step = array.optInt(index)
-                if (step > 0) step else null
-            }
-        }.getOrDefault(emptyList())
-    }
-
-    private fun removePendingProgressBatch(courseId: String, stepNumbers: Set<Int>) {
-        if (stepNumbers.isEmpty()) return
-        val remaining = getPendingProgress(courseId).filterNot { stepNumbers.contains(it) }
-        if (remaining.isEmpty()) {
-            pendingProgressPrefs.edit { remove(progressKey(courseId)) }
-            return
-        }
-        val array = JSONArray()
-        remaining.forEach { array.put(it) }
-        pendingProgressPrefs.edit { putString(progressKey(courseId), array.toString()) }
-    }
-
-    private fun progressKey(courseId: String): String = "course_progress_$courseId"
-
-    private fun bindAttachments(
-        resources: List<DashboardCoursePageFragment.CourseItem.LessonResource>,
-        survey: SurveyDocument?,
-        exam: SurveyDocument?,
-        container: LinearLayout,
-        titleView: TextView,
-        listContainer: LinearLayout
-    ) {
-        resetAttachmentState(listContainer)
-
-        val videoResources = mutableListOf<DashboardCoursePageFragment.CourseItem.LessonResource>()
-        val imageResources = mutableListOf<DashboardCoursePageFragment.CourseItem.LessonResource>()
-        val displayResources = mutableListOf<DashboardCoursePageFragment.CourseItem.LessonResource>()
-
-        resources.forEach { resource ->
-            val mediaType = resource.mediaType.lowercase(Locale.ROOT)
-            val isVideo = mediaType.contains("video")
-            val isImage = mediaType.contains("image")
-            val isPdf = mediaType.contains("pdf")
-            val isAudio = mediaType.contains("audio")
-
-            if (isVideo) videoResources.add(resource)
-            if (isImage) imageResources.add(resource)
-            if (isVideo || isImage || isPdf || isAudio) {
-                displayResources.add(resource)
-            }
-        }
-
-        val hasSurvey = survey?.questions?.isNotEmpty() == true
-        val hasExam = exam?.questions?.isNotEmpty() == true
-        if (displayResources.isEmpty() && !hasSurvey && !hasExam) {
-            container.visibility = View.GONE
-            return
-        }
-        container.visibility = View.VISIBLE
-        titleView.visibility = View.VISIBLE
-        val inflater = layoutInflater
-        prepareVideoPlaylist(videoResources)
-        if (videoResources.isNotEmpty() && currentPlaylistUrls.isEmpty()) {
-            container.visibility = View.GONE
-            Toast.makeText(this, getString(R.string.course_wizard_play_error), Toast.LENGTH_SHORT)
-                .show()
-            return
-        }
-        populateAttachmentList(
-            AttachmentConfig(
-                survey = survey,
-                hasSurvey = hasSurvey,
-                exam = exam,
-                hasExam = hasExam,
-                displayResources = displayResources,
-                imageResources = imageResources,
-                inflater = inflater,
-                listContainer = listContainer
-            )
-        )
-    }
-
-    private fun resetAttachmentState(listContainer: LinearLayout) {
-        listContainer.removeAllViews()
-        playlistIndexByResourceId.clear()
-        currentPlaylistUrls.clear()
-        lastPlaybackIndex = 0
-        lastPlaybackPositionMs = 0L
-        releaseAudioPlayers()
-    }
-
-    private fun prepareVideoPlaylist(videoResources: List<DashboardCoursePageFragment.CourseItem.LessonResource>) {
-        videoResources.forEachIndexed { index, resource ->
-            val resolvedUrl = buildResourceUrl(resource)
-            if (resolvedUrl != null) {
-                playlistIndexByResourceId[resource.id] = index
-                currentPlaylistUrls.add(resolvedUrl)
-            }
-        }
-    }
-
-    private data class AttachmentConfig(
-        val survey: SurveyDocument?,
-        val hasSurvey: Boolean,
-        val exam: SurveyDocument?,
-        val hasExam: Boolean,
-        val displayResources: List<DashboardCoursePageFragment.CourseItem.LessonResource>,
-        val imageResources: List<DashboardCoursePageFragment.CourseItem.LessonResource>,
-        val inflater: android.view.LayoutInflater,
-        val listContainer: LinearLayout
-    )
-
-    private fun populateAttachmentList(config: AttachmentConfig) {
-        if (config.survey != null && config.hasSurvey) {
-            bindSurveyAttachment(config.survey, config.inflater, config.listContainer)
-        }
-        if (config.exam != null && config.hasExam) {
-            bindExamAttachment(config.exam, config.inflater, config.listContainer)
-        }
-        config.displayResources.forEach { resource ->
-            bindResourceAttachment(resource, config.imageResources, config.inflater, config.listContainer)
-        }
-    }
-
-    private fun bindSurveyAttachment(
-        survey: SurveyDocument,
-        inflater: android.view.LayoutInflater,
-        listContainer: LinearLayout
-    ) {
-        val itemView = inflater.inflate(
-            R.layout.item_course_wizard_attachment,
-            listContainer,
-            false
-        )
-        val titleText: TextView = itemView.findViewById(R.id.courseWizardAttachmentTitle)
-        val subtitle: TextView = itemView.findViewById(R.id.courseWizardAttachmentSubtitle)
-        val iconView: android.widget.ImageView =
-            itemView.findViewById(R.id.courseWizardAttachmentIcon)
-        val playButton: ImageButton = itemView.findViewById(R.id.courseWizardAttachmentPlay)
-        titleText.text = survey.name.orEmpty().ifBlank {
-            getString(R.string.dashboard_surveys_title)
-        }
-        subtitle.text = getString(R.string.dashboard_surveys_title)
-        iconView.setImageResource(R.drawable.ic_surveys_24)
-        iconView.contentDescription = getString(R.string.dashboard_surveys_title)
-        playButton.contentDescription = getString(R.string.course_wizard_open_survey)
-        val openSurvey = {
-            pendingRequiredStepIndex = currentIndex
-            SurveyWizardActivity.newIntent(
-                this,
-                survey,
-                survey.teamId,
-                null,
-                courseId
-            ).also { requiredStepLauncher.launch(it) }
-        }
-        itemView.setOnClickListener { openSurvey() }
-        playButton.setOnClickListener { openSurvey() }
-        listContainer.addView(itemView)
-    }
-
-    private fun bindExamAttachment(
-        exam: SurveyDocument,
-        inflater: android.view.LayoutInflater,
-        listContainer: LinearLayout
-    ) {
-        val itemView = inflater.inflate(
-            R.layout.item_course_wizard_attachment,
-            listContainer,
-            false
-        )
-        val titleText: TextView = itemView.findViewById(R.id.courseWizardAttachmentTitle)
-        val subtitle: TextView = itemView.findViewById(R.id.courseWizardAttachmentSubtitle)
-        val iconView: android.widget.ImageView =
-            itemView.findViewById(R.id.courseWizardAttachmentIcon)
-        val playButton: ImageButton = itemView.findViewById(R.id.courseWizardAttachmentPlay)
-        titleText.text = exam.name.orEmpty().ifBlank {
-            getString(R.string.dashboard_exam_title)
-        }
-        subtitle.text = getString(R.string.dashboard_exam_title)
-        iconView.setImageResource(R.drawable.ic_course_step_exam_24)
-        iconView.contentDescription = getString(R.string.dashboard_exam_title)
-        playButton.contentDescription = getString(R.string.course_wizard_open_exam)
-        val openExam = {
-            pendingRequiredStepIndex = currentIndex
-            SurveyWizardActivity.newIntent(
-                this,
-                exam,
-                exam.teamId,
-                null,
-                courseId,
-                isExam = true
-            ).also { requiredStepLauncher.launch(it) }
-        }
-        itemView.setOnClickListener { openExam() }
-        playButton.setOnClickListener { openExam() }
-        listContainer.addView(itemView)
-    }
-
-    private fun bindResourceAttachment(
-        resource: DashboardCoursePageFragment.CourseItem.LessonResource,
-        imageResources: List<DashboardCoursePageFragment.CourseItem.LessonResource>,
-        inflater: android.view.LayoutInflater,
-        listContainer: LinearLayout
-    ) {
-        val mediaTypeLower = resource.mediaType.lowercase(Locale.ROOT)
-        val isAudio = mediaTypeLower.contains("audio")
-        val itemView = if (isAudio) {
-            inflater.inflate(
-                R.layout.item_course_wizard_audio_attachment,
-                listContainer,
-                false
-            )
-        } else {
-            inflater.inflate(
-                R.layout.item_course_wizard_attachment,
-                listContainer,
-                false
-            )
-        }
-        val titleText: TextView = itemView.findViewById(R.id.courseWizardAttachmentTitle)
-        val subtitle: TextView = itemView.findViewById(R.id.courseWizardAttachmentSubtitle)
-        val iconView: android.widget.ImageView =
-            itemView.findViewById(R.id.courseWizardAttachmentIcon)
-        titleText.text = resource.filename
-        subtitle.text = resource.mediaType
-        val isVideo = mediaTypeLower.contains("video")
-        val isImage = mediaTypeLower.contains("image")
-        when {
-            isVideo -> {
-                val playButton: ImageButton = itemView.findViewById(R.id.courseWizardAttachmentPlay)
-                iconView.setImageResource(R.drawable.ic_course_resource_video_24)
-                iconView.contentDescription = getString(R.string.course_wizard_attachment_video)
-                playButton.contentDescription = getString(R.string.course_wizard_play_video)
-            }
-            isImage -> {
-                val playButton: ImageButton = itemView.findViewById(R.id.courseWizardAttachmentPlay)
-                iconView.setImageResource(R.drawable.ic_course_step_image_24)
-                iconView.contentDescription = getString(R.string.course_wizard_attachment_image)
-                playButton.contentDescription = getString(R.string.course_wizard_open_image)
-            }
-            isAudio -> {
-                iconView.setImageResource(R.drawable.ic_course_step_audio_24)
-                iconView.contentDescription = getString(R.string.course_wizard_attachment_audio)
-                bindAudioPlayer(itemView, resource)
-            }
-            else -> {
-                val playButton: ImageButton = itemView.findViewById(R.id.courseWizardAttachmentPlay)
-                iconView.setImageResource(R.drawable.ic_course_step_pdf_24)
-                iconView.contentDescription = getString(R.string.course_wizard_attachment_pdf)
-                playButton.contentDescription = getString(R.string.course_wizard_open_pdf)
-            }
-        }
-        val openResource = {
-            if (isVideo) {
-                if (playlistIndexByResourceId.containsKey(resource.id)) {
-                    selectResource(resource)
-                } else {
-                    Toast.makeText(
-                        this,
-                        getString(R.string.course_wizard_play_error),
-                        Toast.LENGTH_SHORT
-                    ).show()
-                }
-            } else if (isImage) {
-                openImageResource(resource, imageResources)
-            } else {
-                openPdfResource(resource)
-            }
-        }
-        if (!isAudio) {
-            val playButton: ImageButton = itemView.findViewById(R.id.courseWizardAttachmentPlay)
-            itemView.setOnClickListener { openResource() }
-            playButton.setOnClickListener { openResource() }
-        }
-        listContainer.addView(itemView)
-    }
-
-    private fun bindAudioPlayer(
-        itemView: View,
-        resource: DashboardCoursePageFragment.CourseItem.LessonResource
-    ) {
-        val url = buildResourceUrl(resource)
-        if (url.isNullOrBlank()) {
-            Toast.makeText(this, getString(R.string.course_wizard_play_error), Toast.LENGTH_SHORT)
-                .show()
-            return
-        }
-        val authHeader = if (baseUrl?.startsWith("https://", ignoreCase = true) == true) {
-            credentials?.let { Credentials.basic(it.username, it.password) }
-        } else {
-            null
-        }
-        val playerView: PlayerView = itemView.findViewById(R.id.courseWizardAudioPlayer)
-        val player = ExoPlayer.Builder(this)
-            .setMediaSourceFactory(DefaultMediaSourceFactory(buildAudioDataSourceFactory(authHeader)))
-            .build()
-        playerView.player = player
-        player.setMediaItem(MediaItem.fromUri(url))
-        player.prepare()
-        player.playWhenReady = false
-        audioPlayers.add(player)
-    }
-
-    private fun buildAudioDataSourceFactory(authorizationHeader: String?): DataSource.Factory {
-        val httpFactory = DefaultHttpDataSource.Factory().setAllowCrossProtocolRedirects(true)
-        authorizationHeader?.let { httpFactory.setDefaultRequestProperties(mapOf("Authorization" to it)) }
-        return DefaultDataSource.Factory(this, httpFactory)
-    }
-
-    private fun releaseAudioPlayers() {
-        audioPlayers.forEach { it.release() }
-        audioPlayers.clear()
-    }
-
-    private fun selectResource(resource: DashboardCoursePageFragment.CourseItem.LessonResource) {
-        val targetIndex = playlistIndexByResourceId[resource.id] ?: return
-        launchFullscreenPlayer(targetIndex)
-    }
-
-    private fun launchFullscreenPlayer(startIndex: Int) {
-        if (currentPlaylistUrls.isEmpty()) {
-            Toast.makeText(this, getString(R.string.course_wizard_play_error), Toast.LENGTH_SHORT)
-                .show()
-            return
-        }
-        val startPosition = if (startIndex == lastPlaybackIndex) {
-            lastPlaybackPositionMs
-        } else {
-            0L
-        }
-        val authHeader = if (baseUrl?.startsWith("https://", ignoreCase = true) == true) {
-            credentials?.let { Credentials.basic(it.username, it.password) }
-        } else {
-            null
-        }
-        val intent = FullscreenPlayerActivity.createIntent(
-            context = this,
-            mediaUrls = ArrayList(currentPlaylistUrls),
-            startIndex = startIndex,
-            startPositionMs = startPosition,
-            authorizationHeader = authHeader
-        )
-        fullscreenLauncher.launch(intent)
-    }
-
-    private fun openPdfResource(resource: DashboardCoursePageFragment.CourseItem.LessonResource) {
-        val url = buildResourceUrl(resource)
-        if (url.isNullOrBlank()) {
-            Toast.makeText(this, getString(R.string.course_wizard_play_error), Toast.LENGTH_SHORT)
-                .show()
-            return
-        }
-        val authHeader = if (baseUrl?.startsWith("https://", ignoreCase = true) == true) {
-            credentials?.let { Credentials.basic(it.username, it.password) }
-        } else {
-            null
-        }
-        val intent = FullscreenPdfActivity.createIntent(this, url, authHeader)
-        startActivity(intent)
-    }
-
-    private fun openImageResource(
-        resource: DashboardCoursePageFragment.CourseItem.LessonResource,
-        imageResources: List<DashboardCoursePageFragment.CourseItem.LessonResource>
-    ) {
-        val validResources = imageResources.mapNotNull { image ->
-            buildResourcePath(image)?.let { path -> image to path }
-        }
-        if (validResources.isEmpty()) {
-            Toast.makeText(this, getString(R.string.course_wizard_play_error), Toast.LENGTH_SHORT)
-                .show()
-            return
-        }
-        val imagePaths = validResources.map { it.second }
-        val startIndex = validResources.indexOfFirst {
-            it.first.id == resource.id && it.first.filename == resource.filename
-        }
-            .coerceAtLeast(0)
-        val intent = Intent(this, DashboardImagePreviewActivity::class.java).apply {
-            putStringArrayListExtra(
-                DashboardImagePreviewActivity.EXTRA_IMAGE_PATHS,
-                ArrayList(imagePaths)
-            )
-            putExtra(DashboardImagePreviewActivity.EXTRA_START_INDEX, startIndex)
-        }
-        startActivity(intent)
-    }
-
-    private fun buildResourceUrl(resource: DashboardCoursePageFragment.CourseItem.LessonResource): String? {
-        val safeCourseId = courseId?.takeIf { it.isNotBlank() } ?: return null
-        val localFile = OfflineCourseStorage.findExistingResourceFile(
-            this,
-            safeCourseId,
-            resource.id,
-            resource.filename
-        )
-        if (localFile?.exists() == true) {
-            return localFile.toURI().toString()
-        }
-        val normalizedBase = baseUrl?.trim()?.trimEnd('/')?.takeIf { it.isNotEmpty() } ?: return null
-        val parsed = normalizedBase.toUri()
-        val scheme = parsed.scheme ?: return null
-        val authority = parsed.encodedAuthority ?: return null
-        return parsed.buildUpon()
-            .scheme(scheme)
-            .encodedAuthority(authority)
-            .appendPath("db")
-            .appendPath("resources")
-            .appendPath(resource.id)
-            .appendPath(resource.filename)
-            .build()
-            .toString()
-    }
-
-    private fun buildResourcePath(resource: DashboardCoursePageFragment.CourseItem.LessonResource): String? {
-        val safeCourseId = courseId?.takeIf { it.isNotBlank() }
-        if (safeCourseId != null) {
-            val localFile = OfflineCourseStorage.findExistingResourceFile(
-                this,
-                safeCourseId,
-                resource.id,
-                resource.filename
-            )
-            if (localFile?.exists() == true) {
-                return localFile.toURI().toString()
-            }
-        }
-        val resourceId = resource.id.trim().takeIf { it.isNotEmpty() } ?: return null
-        val filename = resource.filename.trim().takeIf { it.isNotEmpty() } ?: return null
-        return "resources/$resourceId/$filename"
-    }
-
-    private fun resolveOfflineMarkdownImages(markdown: String): String {
-        return MarkdownUtils.resolveOfflineMarkdownImages(this, markdown, courseId, baseUrl)
-    }
-
-    data class StepDisplay(
-        val title: String,
-        val description: String,
-        val resources: List<DashboardCoursePageFragment.CourseItem.LessonResource>,
-        val survey: SurveyDocument? = null,
-        val exam: SurveyDocument? = null
-    )
-
     companion object {
         private const val PREF_LEGACY_PENDING_COURSE_PROGRESS = "pref_pending_course_progress"
         private const val PREF_ENCRYPTED_PENDING_COURSE_PROGRESS = "encrypted_pending_course_progress"
@@ -904,7 +270,7 @@ class CourseWizardActivity : BaseActivity() {
             context: Context,
             courseId: String,
             courseTitle: String,
-            steps: List<DashboardCoursePageFragment.CourseItem.LessonStep>,
+            steps: List<CourseItem.LessonStep>,
             startStep: Int
         ): Intent {
             return Intent(context, CourseWizardActivity::class.java).apply {
@@ -919,7 +285,7 @@ class CourseWizardActivity : BaseActivity() {
             context: Context,
             courseId: String,
             courseTitle: String,
-            steps: List<DashboardCoursePageFragment.CourseItem.LessonStep>,
+            steps: List<CourseItem.LessonStep>,
             startStep: Int
         ) {
             val intent = createIntent(context, courseId, courseTitle, steps, startStep)
